+++ source/dotnet/Services/Domain/Actor/IActorRepository.cs
@@ -12,6 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Domain.BatchActor;
+using Energinet.DataHub.Wholesale.Domain.GridAreaAggregate;
+using Energinet.DataHub.Wholesale.Domain.ProcessStepResultAggregate;
 
-public record Actor(string Gln);
+namespace Energinet.DataHub.Wholesale.Domain.Actor;
+
+public interface IActorRepository
+{
+    Task<Domain.Actor.Actor[]> GetAsync(Guid batchId, GridAreaCode gridAreaCode, TimeSeriesType timeSeriesType, MarketRoleType marketRoleType);
+}
