+++ source/dotnet/Services/Domain/Actor/MarketRoleType.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Domain.BatchActor;
+namespace Energinet.DataHub.Wholesale.Domain.Actor;
 
 public enum MarketRoleType
 {
