+++ source/dotnet/Services/Infrastructure/BatchActor/ActorRepository.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using Azure.Storage.Files.DataLake;
-using Energinet.DataHub.Wholesale.Domain.BatchActor;
+using Energinet.DataHub.Wholesale.Domain.Actor;
 using Energinet.DataHub.Wholesale.Domain.GridAreaAggregate;
 using Energinet.DataHub.Wholesale.Domain.ProcessStepResultAggregate;
 using Energinet.DataHub.Wholesale.Infrastructure.Persistence.DataLake;
@@ -33,7 +33,7 @@ public class ActorRepository : DataLakeRepositoryBase, IActorRepository
         _dataLakeTypeFactory = dataLakeTypeFactory;
     }
 
-    public async Task<Domain.BatchActor.Actor[]> GetAsync(
+    public async Task<Domain.Actor.Actor[]> GetAsync(
         Guid batchId,
         GridAreaCode gridAreaCode,
         TimeSeriesType timeSeriesType,
@@ -57,8 +57,8 @@ public class ActorRepository : DataLakeRepositoryBase, IActorRepository
         return ($"calculation-output/batch_id={batchId}/actors/grid_area={gridAreaCode.Code}/time_series_type={TimeSeriesTypeMapper.Map(timeSeriesType)}/market_role={MarketRoleTypeMapper.Map(marketRoleType)}/", ".json");
     }
 
-    private static Domain.BatchActor.Actor[] MapToBatchActor(IEnumerable<Actor> actors)
+    private static Domain.Actor.Actor[] MapToBatchActor(IEnumerable<Actor> actors)
     {
-        return actors.Select(actor => new Domain.BatchActor.Actor(actor.gln)).ToArray();
+        return actors.Select(actor => new Domain.Actor.Actor(actor.gln)).ToArray();
     }
 }
