+++ source/dotnet/Services/Infrastructure/BatchActor/MarketRoleTypeMapper.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Domain.BatchActor;
+using Energinet.DataHub.Wholesale.Domain.Actor;
 
 namespace Energinet.DataHub.Wholesale.Infrastructure.BatchActor;
