+++ source/dotnet/Services/UnitTests/Infrastructure/BatchActor/MarketRoleTypeMapper.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Domain.BatchActor;
+using Energinet.DataHub.Wholesale.Domain.Actor;
 using Energinet.DataHub.Wholesale.Infrastructure.BatchActor;
 using FluentAssertions;
 using Xunit;
